Add files via upload
This is developed on around Nov. 2017, when I was a junior student. This application was aimed at achieving basic survices of a GISystem, such as painting a map, searching (map to data & data to map) , requesting for names and coordinates. In logic aspect, the application tried to establish grid index in case a larger search burden from server. Twelve rectangle grids were created to build this index. No exterior libraries seemed to be added in to this program. The map painting module was achieved by reading .txt file and painted by QPainter.